backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
starlette>=0.40.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1,<6.0.0
redis>=5.0.0

langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-ollama>=0.2.0

psycopg2-binary>=2.9.9
pgvector>=0.3.0

python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

pymupdf>=1.24.0
pillow>=10.3.0
jieba>=0.42.1
sentence-transformers>=3.0.0
numpy>=1.26.0
matplotlib>=3.9.0
